Add CareKitEssentialChartable protocol
